コードをみやすく、テストもしやすくするために、１クラス１ファイルに分割
